inference_api/pyproject.toml
[tool.poetry]
name = "inference_api"
version = "0.1.0"
description = ""
authors = ["Nicholas Shook <nick@neonlaw.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
pytesseract = "^0.3.3"
pillow = "^7.1.1"
numpy = "^1.18.2"
opencv-python = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
